refactor[#12]: rename SearchView to SearchPlanView
