+++ src/router/index.ts
@@ -1,6 +1,6 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import HomeView from '../views/HomeView.vue'
-import SearchView from '@/views/search/SearchView.vue'
+import SearchPlanView from '@/views/search-plan/SearchPlanView.vue'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -21,7 +21,7 @@ const router = createRouter({
     {
       path: '/search',
       name: 'search',
-      component: SearchView,
+      component: SearchPlanView,
     },
   ],
 })
